requirements.txt
Flask>=3.0,<3.2
flask-cors==4.0.0
gunicorn==21.2.0
openai==0.28.0
cohere==5.17.0
requests>=2.31
Flask
python-dotenv
Pillow
reportlab